Move set_colour to cli_logger
